+++ circularbuffer.h
@@ -7,12 +7,12 @@
 namespace emb {
 
 
-template <typename T, int Capacity>
+template <typename T, size_t Capacity>
 class circular_buffer {
 private:
     T _data[Capacity];
-    int _front;
-    int _back;
+    size_t _front;
+    size_t _back;
     bool _full;
 public:
     circular_buffer()
@@ -29,10 +29,10 @@ public:
 
     bool empty() const { return (!_full && (_front == _back)); }
     bool full() const { return _full; }
-    int capacity() const { return Capacity; }
+    size_t capacity() const { return Capacity; }
 
-    int size() const {
-        int size = Capacity;
+    size_t size() const {
+        size_t size = Capacity;
         if (!_full) {
             if (_back >= _front) {
                 size = _back - _front;
@@ -73,7 +73,7 @@ public:
     const T* end() const { return _data + Capacity; }
 
     void fill(const T& value) {
-        for (int i = 0; i < Capacity; ++i) {
+        for (size_t i = 0; i < Capacity; ++i) {
             _data[i] = value;
         }
     }
