+++ core/c28x.h
@@ -15,7 +15,7 @@ namespace c28x {
 template <typename T>
 void from_bytes(T& dest, const uint8_t* src) {
     uint16_t c28_bytes[sizeof(T)];
-    for (int i = 0; i < sizeof(T); ++i) {
+    for (size_t i = 0; i < sizeof(T); ++i) {
         c28_bytes[i] = src[2*i] | src[2*i+1] << 8;
     }
     memcpy (&dest, &c28_bytes, sizeof(T));
@@ -26,7 +26,7 @@ template <typename T>
 void to_bytes(uint8_t* dest, const T& src) {
     uint16_t c28_bytes[sizeof(T)];
     memcpy(&c28_bytes, &src, sizeof(T));
-    for (int i = 0; i < sizeof(T); ++i) {
+    for (size_t i = 0; i < sizeof(T); ++i) {
         dest[2*i] = c28_bytes[i] & 0x00FF;
         dest[2*i+1] = c28_bytes[i] >> 8;
     }
@@ -41,7 +41,7 @@ bool are_equal(const T& obj1, const T& obj2) {
     to_bytes<T>(obj1_bytes, obj1);
     to_bytes<T>(obj2_bytes, obj2);
 
-    for(int i = 0; i < sizeof(T)*2; ++i) {
+    for(size_t i = 0; i < sizeof(T)*2; ++i) {
         if (obj1_bytes[i] != obj2_bytes[i]) {
             return false;
         }
